optional-requirements.txt
cupy
pillow
astropy
astropy
photutils
matplotlib
skimage